// src/models/predict-example-response.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v4-mini";
import { remap as remap$ } from "../lib/primitives.js";
import { safeParse } from "../lib/schemas.js";
import { Result as SafeParseResult } from "../types/fp.js";
import * as types from "../types/primitives.js";
import { SDKValidationError } from "./errors/sdk-validation-error.js";

export type PredictExampleResponse = {
  exampleId: string;
  predictionId: string;
  output?: { [k: string]: any } | null | undefined;
  reasoning?: string | null | undefined;
  messages?: Array<{ [k: string]: any }> | null | undefined;
};

/** @internal */
export const PredictExampleResponse$inboundSchema: z.ZodMiniType<
  PredictExampleResponse,
  unknown
> = z.pipe(
  z.object({
    example_id: types.string(),
    prediction_id: types.string(),
    output: z.optional(z.nullable(z.record(z.string(), z.any()))),
    reasoning: z.optional(z.nullable(types.string())),
    messages: z.optional(z.nullable(z.array(z.record(z.string(), z.any())))),
  }),
  z.transform((v) => {
    return remap$(v, {
      "example_id": "exampleId",
      "prediction_id": "predictionId",
    });
  }),
);

export function predictExampleResponseFromJSON(
  jsonString: string,
): SafeParseResult<PredictExampleResponse, SDKValidationError> {
  return safeParse(
    jsonString,
    (x) => PredictExampleResponse$inboundSchema.parse(JSON.parse(x)),
    `Failed to parse 'PredictExampleResponse' from JSON`,
  );
}